# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Empresta API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/convenios": {
            "get": {
                "description": "Retorna um objeto JSON com todos os convênios.",
                "operationId": "App\\Http\\Controllers\\Api\\ConvenioController::exportarConvenios",
                "responses": {
                    "200": {
                        "description": "OK."
                    },
                    "400": {
                        "description": "Houve algo de errado na requisição."
                    },
                    "500": {
                        "description": "Erro interno da aplicação."
                    }
                }
            }
        },
        "/credito/simular": {
            "post": {
                "description": "Simulação de crédito para o cliente.",
                "operationId": "App\\Http\\Controllers\\Api\\CreditoController::simular",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "valor_emprestimo"
                                ],
                                "properties": {
                                    "valor_emprestimo": {
                                        "type": "float",
                                        "example": "1500"
                                    },
                                    "instituicoes": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "BMG"
                                        }
                                    },
                                    "convenios": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "INSS"
                                        }
                                    },
                                    "parcelas": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": "60"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK."
                    },
                    "400": {
                        "description": "Houve algo de errado na requisição."
                    },
                    "500": {
                        "description": "Erro interno da aplicação."
                    }
                }
            }
        },
        "/instituicoes": {
            "get": {
                "description": "Retorna um objeto JSON com todos as instituições.",
                "operationId": "App\\Http\\Controllers\\Api\\InstituicaoController::exportarInstituicoes",
                "responses": {
                    "200": {
                        "description": "OK."
                    },
                    "400": {
                        "description": "Houve algo de errado na requisição."
                    },
                    "500": {
                        "description": "Erro interno da aplicação."
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}